requirements.txt
solana==0.34.2
solders==0.21.0
supabase==2.4.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
base58==2.1.1 